refactor: extract camera widgets into lib/screen/camera/ (Phase 1)
- Split camera_screen.dart (1966 -> 1411 lines) by extracting 8 private
  widgets/painters into separate files under lib/screen/camera/widgets/:
  GlassIconButton, ZoomPill, TopCameraBar, PortraitTopBar,
  BottomCameraControls, CaptureButton, ModeSwitcher, RoiPainter,
  ThirdsGridPainter.
- Unify duplicate ShootingMode enum (previously in both coaching_result.dart
  and camera_screen.dart) into lib/screen/camera/shooting_mode.dart.
- No behavior change; flutter analyze passes with no new errors.
