+++ lib/screen/camera/widgets/thirds_grid_painter.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+/// 3분할선만 그리는 기본 페인터.
+///
+/// **Note**: Phase 2에서 `CompositionRule` 주입형 `CompositionGridPainter`로
+/// 교체될 임시 위치. 그 전까지는 기존 동작을 유지한다.
+class ThirdsGridPainter extends CustomPainter {
+  const ThirdsGridPainter();
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final paint = Paint()
+      ..color = const Color(0x33FFFFFF)
+      ..strokeWidth = 1;
+
+    final dx1 = size.width / 3;
+    final dx2 = size.width * 2 / 3;
+    final dy1 = size.height / 3;
+    final dy2 = size.height * 2 / 3;
+
+    canvas.drawLine(Offset(dx1, 0), Offset(dx1, size.height), paint);
+    canvas.drawLine(Offset(dx2, 0), Offset(dx2, size.height), paint);
+    canvas.drawLine(Offset(0, dy1), Offset(size.width, dy1), paint);
+    canvas.drawLine(Offset(0, dy2), Offset(size.width, dy2), paint);
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
+}
