+++ lib/screen/camera/widgets/zoom_pill.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+
+/// 카메라 줌 프리셋 선택용 원형 버튼 (1x, 2x 등).
+class ZoomPill extends StatelessWidget {
+  final String label;
+  final bool selected;
+  final VoidCallback onTap;
+
+  const ZoomPill({
+    super.key,
+    required this.label,
+    required this.selected,
+    required this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Container(
+        width: selected ? 40 : 34,
+        height: selected ? 32 : 28,
+        alignment: Alignment.center,
+        decoration: BoxDecoration(
+          color: selected ? Colors.white : const Color(0x1AFFFFFF),
+          borderRadius: BorderRadius.circular(999),
+        ),
+        child: Text(
+          label,
+          style: TextStyle(
+            color: selected ? const Color(0xFF333333) : Colors.white,
+            fontSize: selected ? 11 : 10,
+            fontWeight: FontWeight.w700,
+          ),
+        ),
+      ),
+    );
+  }
+}
